// ssr/models/pokemon.model.ts
export type PokemonType =
  | "Normal"
  | "Fire"
  | "Water"
  | "Electric"
  | "Grass"
  | "Ice"
  | "Fighting"
  | "Poison"
  | "Ground"
  | "Flying"
  | "Psychic"
  | "Bug"
  | "Rock"
  | "Ghost"
  | "Dragon"
  | "Dark"
  | "Steel"
  | "Fairy";

export type Pokemon = {
  id: number;
  image?: string;
  name: {
    english: string;
    japanese: string;
    chinese: string;
    french: string;
  };
  type: PokemonType[];
  base: {
    HP: number;
    Attack: number;
    Defense: number;
    "Sp.Attack": number;
    "Sp.Defense": number;
    Speed: number;
  };
};